requirements.txt
streamlit
httpx
beautifulsoup4
lxml
google-generativeai
pydantic>=2.0
python-dotenv
